irqchip: meson: add support for gpio interrupt controller
Add support for the interrupt gpio controller found on Amlogic's meson
SoC family.

Unlike what the IP name suggest, it is not directly linked to the gpio
subsystem. It is actually independent from the gpio subsystem. This
controller is able  to spy on the SoC pad. For this purpose, it can mux
and filter (edge or level and polarity) *almost* any pad of the SoC to one
of the 8 GIC's interrupt lines it owns. There is a few pins which are not
available on the controller input, especially on meson8b

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>