clk: check ops pointer on clock register
Nothing really prevent a provider from (trying to) registering a clock
without providing the clock ops structure.

We do check the individual fields before using them, but not the structure
pointer itself. This may have the usual nasty consequences when the
pointer is dereferenced, mostly likely when checking one the field during
the initialization.

This is fixed by returning an error on clock register if the ops pointer
is NULL

Fixes: b2476490ef11 ("clk: introduce the common clock framework")
Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>