regmap: add regmap_might_sleep()
Add regmap_might_sleep() to let users of regmap query its sleepiness.
This is useful for users which need to act differently if the register
access might sleep. This is the case in the common clock framework where,
if the register access might sleep, the job usually done in enable() needs
to be done in prepare() instead.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>