clk: gate: fixup tentative
This is a tentative to silently migrate clock users to my new scheme.
Next step would be have been to changes the ops to use clk_hw_read/write.
But we can't change the ops in this way, some other driver driver a calling
those ops and would be broken as a result ... moving to ng approach

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>