clk: meson: rework around meson8b cpu clock
Instead of migrating meson cpu_clk to clk_regmap, like the other meson
clock drivers, we take advantage of the massive rework to get rid of it
completly.

As it was pointed out in different FIXME comment in the code, the cpu_clk
should be modeled with dividers and muxes it is made of, instead of one
big composite clock.

The other issue pointed out in the FIXME note, around cpu notifier,
remains unsolved, until CCR comes up.

Signed-off-by: Jerome Brunet <jbrunet@baylibre.com>